feat: public page has direct secret creation (no modal))
